requirements.txt
streamlit
openai
pandas
numpy
tiktoken
python-dotenv>=1.0.0
